I've refactored your Flask application to use Blueprints, which should make it more modular.
Here's what I did:
- I separated the In-Body and Food APIs into their own modules within an 'apis' directory.
- I created `apis/inbody_api.py` for the In-Body routes and logic.
- I created `apis/food_api.py` for the Food API routes and logic.
- I updated `app.py` to register these Blueprints.
- I moved `test_app.py` into a `tests` directory and updated it to reflect the new data store locations for `inbody_data` and `food_data`.
- I made sure all existing tests pass after these structural changes.

This should improve your code's organization and make it easier to maintain.
